[refactor] 프로젝트 추가 페이지 아토믹 컴포넌트 적용
- 프로젝트 추가 페이지 아토믹 컴포넌트 적용
- 기존 presenter 컴포넌트를 header와 section으로 분리
- 현재 container 컴포넌트가 너무 크므로 추후 리팩터링 예정
